[fastfit] move MCA stack view and output buffer to separate files. Temporary fix unit test by conversion to float32.
